CompileReactor - Ability to compile a project and get responses from it Project - Updates for restarting engine, timeout for sockets IProject - Update to get project asset folder. Although, we should shift this to the compile directly here I think.
git-svn-id: svn://107.170.10.188:8080/Semoss/dev@49042 ef391c1c-ee15-45f7-b1d6-dde3d27a2169

Former-commit-id: 7a27dfd47888cb90b4807116d3bfddef991e98ba

+++ src/prerna/ds/py/PyTranslator.java
@@ -10,11 +10,13 @@ import java.util.List;
 import java.util.Map;
 
 import org.apache.commons.io.FileUtils;
+import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
 import prerna.algorithm.api.SemossDataType;
 import prerna.auth.utils.AbstractSecurityUtils;
 import prerna.om.Insight;
+import prerna.project.impl.Project;
 import prerna.tcp.client.ErrorSenderThread;
 import prerna.tcp.client.NativePySocketClient;
 import prerna.util.AssetUtility;
@@ -59,6 +61,7 @@ public class PyTranslator {
 	public PyTranslator() {
 		// startDisruptor();
 		// System.out.println("Py Translator created");
+		this.logger = LogManager.getLogger(Project.class);
 	}
 
 	public SemossDataType convertDataType(String pDataType) {
@@ -573,12 +576,15 @@ public class PyTranslator {
 				throw new IllegalArgumentException("Failed to run Py script.");
 			} finally {
 				// Cleanup
-				outputFile.delete();
-				try {
+				if(outputFile.exists())
+				{
+					try {
+					outputFile.delete();
 					this.runScript(removePathVariables);
 					// this.executeEmptyR("gc();"); // Garbage collection
-				} catch (Exception e) {
+					} catch (Exception e) {
 					logger.warn("Unable to cleanup Py.", e);
+					}
 				}
 			}
 		} catch (IOException e) {
